requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi 
# pymongo[srv]==3.6

## -e . it will conver the whole project into package